employee_api/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
python-decouple==3.8
